New add account frame. Small restructuring.
I added a new frame for the Add Account window. In addition, I allowed
both the login and add account windows to open when main is run. Nothing
yet happens for the save button.

In addition, I added an XML directory to store all XML files to be
easily referenced and moved such files there.