Created BankFrame and related forms.
Created and allowed for initialization of the bank frame. This frame
will now open the add account frame and the login frame, as well as make
changes to the JList in the frame with the remove button.

In addition, created xml files for the bank frame and the account frame.
The code for the account frame will come with the next commit.