Adding to report
I’m not sure as to what happened, but the app is not working. keep
getting a “AWT-EventQueue-0" java.lang.NullPointerException” error
message and I am not sure as to what is causing the problem.

The report is complete. I just need images, but since the app is no
long working…idk how to obtain those images. It deals with the GUI
which I am not familiar with…